+++ src/main/java/nextstep/subway/domain/SubwayDurationMap.java
@@ -1,5 +1,8 @@
 package nextstep.subway.domain;
 
+import nextstep.subway.domain.line.Line;
+import nextstep.subway.domain.line.Section;
+import nextstep.subway.domain.station.Station;
 import org.jgrapht.graph.SimpleDirectedWeightedGraph;
 
 import java.util.List;
