+++ src/main/java/nextstep/subway/domain/fare/discount/Children.java
@@ -1,4 +1,4 @@
-package nextstep.subway.domain;
+package nextstep.subway.domain.fare.discount;
 
 public class Children extends DiscountFarePolicy {
