+++ src/main/java/nextstep/subway/domain/fare/discount/Teenager.java
@@ -1,4 +1,4 @@
-package nextstep.subway.domain;
+package nextstep.subway.domain.fare.discount;
 
 public class Teenager extends DiscountFarePolicy {
