+++ src/main/java/nextstep/subway/domain/fare/distance/Include.java
@@ -1,4 +1,4 @@
-package nextstep.subway.domain;
+package nextstep.subway.domain.fare.distance;
 
 import nextstep.subway.domain.fare.distance.DistanceFarePolicy;
