+++ src/main/java/nextstep/subway/domain/favorite/Favorite.java
@@ -1,4 +1,4 @@
-package nextstep.subway.domain;
+package nextstep.subway.domain.favorite;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
