+++ src/main/java/nextstep/subway/domain/favorite/FavoriteRepository.java
@@ -1,4 +1,4 @@
-package nextstep.subway.domain;
+package nextstep.subway.domain.favorite;
 
 import org.springframework.data.jpa.repository.JpaRepository;
