+++ src/main/java/nextstep/subway/domain/station/StationRepository.java
@@ -1,5 +1,6 @@
-package nextstep.subway.domain;
+package nextstep.subway.domain.station;
 
+import nextstep.subway.domain.station.Station;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface StationRepository extends JpaRepository<Station, Long> {
